# Conflicts:
#	Barotrauma/BarotraumaClient/Source/GameSession/GameModes/Tutorials/BasicTutorial.cs
#	Barotrauma/BarotraumaClient/Source/GameSession/GameModes/Tutorials/ContextualTutorial.cs
#	Barotrauma/BarotraumaClient/Source/Items/Components/Machines/MiniMap.cs
#	Barotrauma/BarotraumaClient/Source/Items/Components/Machines/Sonar.cs
#	Barotrauma/BarotraumaClient/Source/Items/Components/RepairTool.cs
#	Barotrauma/BarotraumaClient/Source/Items/Inventory.cs
#	Barotrauma/BarotraumaShared/Source/Characters/AI/Objectives/AIObjectiveFixLeaks.cs
#	Barotrauma/BarotraumaShared/Source/Characters/AI/Objectives/AIObjectiveRepairItem.cs
#	Barotrauma/BarotraumaShared/Source/Characters/AI/Objectives/AIObjectiveRepairItems.cs
#	Barotrauma/BarotraumaShared/Source/Items/Components/Door.cs
#	Barotrauma/BarotraumaShared/Source/Items/Components/Machines/Engine.cs
#	Barotrauma/BarotraumaShared/Source/Items/Components/Machines/Fabricator.cs
#	Barotrauma/BarotraumaShared/Source/Items/Components/Machines/MiniMap.cs
#	Barotrauma/BarotraumaShared/Source/Items/Components/Machines/OxygenGenerator.cs
#	Barotrauma/BarotraumaShared/Source/Items/Components/Machines/Pump.cs
#	Barotrauma/BarotraumaShared/Source/Items/Components/Machines/Reactor.cs
#	Barotrauma/BarotraumaShared/Source/Items/Components/Power/PowerTransfer.cs
#	Barotrauma/BarotraumaShared/Source/Items/Components/Repairable.cs

+++ Barotrauma/BarotraumaServer/Source/Characters/CharacterNetworking.cs
@@ -472,7 +472,7 @@ namespace Barotrauma
                 msg.Write(false);
             }
 
-            msg.Write(TeamID);
+            msg.Write((byte)TeamID);
             msg.Write(this is AICharacter);
             info.ServerWrite(msg);
